TAC-3940 | Create service-specific server list classes.
We can reduce complexity in the Environment class by shifting some of its
responsibility to service-specific list classes. This PR adds the skeleton
for those classes, with an example BalancerServerList::getActiveBalancers()
method.

Additionally this PR fixes an issue with Server::setServices() to allow
services which are an empty array to be set (common in Cloud API data).